Changed gains, constants, and made new methods.
I changed gains to make them more accurate and they actually work better. I also made new methods in the Elevator subsystem for setting the mode of the subsystem. Along with this I changed the constants in the elevator subsystem to reflect the Western NE branch of code so it should be more accurate and better. I also made minor changes to make things work better more generally and also I put all the elevator commands in a new folder for organization. Overall, I made alot of progress